refactor rng into separate file
